Fixed issue #5 - keyboard shortcuts not working. There where two problems: first, key bindings needed "focusedView" instead of "view" in their "when" clauses and secondly the "viewItem" had to be completely removed (see https://github.com/microsoft/vscode/issues/72442).
In addition, this commit updates the "Delete" command, to run for all selected files instead of the last selected file. As a result of the implementation, when deleting multiple files the user will be asked whether they are sure for each file separately. To address this problem, a new delete command that unifies existing delete commands should be added.